requirements.txt
torch~=2.2.1
opencv-python~=4.9.0.80
pandas~=2.1.1
numpy~=1.26.1
timm~=0.9.8
natten~=0.14.6
einops~=0.7.0
torchvision~=0.17.1
ptflops~=0.7.1.1
thop~=0.1.1.post2209072238
albumentations~=1.3.1
matplotlib~=3.8.3
scikit-learn~=1.4.1.post1
pillow~=10.2.0
mxnet~=1.6.0
fairscale~=0.4.13
transformers~=4.38.2
torchmetrics~=1.3.2
argparse~=1.4.0
eer~=0.0.2
tensorflow~=2.16.1
six~=1.16.0
imageio~=2.31.6
scipy~=1.12.0